// simulator/src/typechain-types/IAggregatorContract.ts
export declare namespace IAggregatorContract {
  export type AgreementStruct = {
    value: PromiseOrValue<BigNumberish>;
    flexibility: PromiseOrValue<BigNumberish>;
    valuePrice: PromiseOrValue<BigNumberish>;
    flexibilityPrice: PromiseOrValue<BigNumberish>;
    energySource: PromiseOrValue<BigNumberish>;
  };

  export type AgreementStructOutput = [
    BigNumber,
    BigNumber,
    BigNumber,
    BigNumber,
    number
  ] & {
    value: BigNumber;
    flexibility: BigNumber;
    valuePrice: BigNumber;
    flexibilityPrice: BigNumber;
    energySource: number;
  };
}